more on player and added effect implementation

+++ include/models/Effect.hpp
@@ -9,7 +9,7 @@ public:
     virtual void onTurnEnd(Player& p);
     virtual bool isExpired() const = 0;
     virtual bool blocksPayment() const;
-    virtual int modifyPayment(int amount) const;
+    virtual float modifyPayment(float amount) const;
 };
 
 class ShieldEffect : public Effect {
@@ -25,12 +25,12 @@ public:
 
 class DiscountEffect : public Effect {
 private:
-    int percent;
+    float percent;
     int turns_left;
 
 public:
     DiscountEffect(int percent, int duration = 1);
     void onTurnEnd(Player& p) override;
     bool isExpired() const override;
-    int modifyPayment(int amount) const override;
+    float modifyPayment(float amount) const override;
 };
